+++ okitweb/static/okit/model/js/resources/virtual_cloud_network.js
@@ -17,7 +17,7 @@ class VirtualCloudNetwork extends OkitArtifact {
         this.compartment_id = data.parent_id;
         // Generate Cidr
         this.cidr_blocks = [];
-        this.dns_label = '';
+        this.dns_label = this.getOkitJson().metadata.platform === 'pca' ? '' : this.generateDnsLabel()
         // this.dns_label = this.generateDnsLabel();
         this.is_ipv6enabled = false;
         this.ipv6cidr_blocks = [];