+++ okitweb/static/okit/pricing/js/resources/database_system.js
@@ -2,16 +2,16 @@
 ** Copyright (c) 2020, 2022, Oracle and/or its affiliates.
 ** Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 */
-console.debug('Loaded ObjectStorageBucket Pricing Javascript');
+console.debug('Loaded DatabaseSystem Pricing Javascript');
 
 /*
-** Define ObjectStorageBucket Pricing Class
+** Define DatabaseSystem Pricing Class
  */
-class ObjectStorageBucketOciPricing extends OkitOciPricingResource {
+class DatabaseSystemOciPricing extends OkitOciPricingResource {
     constructor(resource, pricing) {
         super(resource, pricing)
     }
-
+    
     getPrice(resource) {
         resource = resource ? resource : this.resource
         const price_per_month = 0
@@ -26,12 +26,12 @@ class ObjectStorageBucketOciPricing extends OkitOciPricingResource {
     }
 }
 
-OkitOciProductPricing.prototype.getObjectStorageBucketPrice = function(resource) {
-    const pricing_resource = new ObjectStorageBucketOciPricing(resource, this)
+OkitOciProductPricing.prototype.getDatabaseSystemPrice = function(resource) {
+    const pricing_resource = new DatabaseSystemOciPricing(resource, this)
     return pricing_resource.getPrice(resource)
 }
 
-OkitOciProductPricing.prototype.getObjectStorageBucketBoM = function(resource) {
-    const pricing_resource = new ObjectStorageBucketOciPricing(resource, this)
+OkitOciProductPricing.prototype.getDatabaseSystemBoM = function(resource) {
+    const pricing_resource = new DatabaseSystemOciPricing(resource, this)
     return pricing_resource.getBoM(resource)
 }