+++ okitweb/static/okit/pricing/js/resources/mysql_database_system.js
@@ -21,7 +21,7 @@ class MysqlDatabaseSystemOciPricing extends OkitOciPricingResource {
 
     getBoM(resource) {
         const resource_name = resource.getArtifactReference()
-        const bom = {skus: {}, price_per_month: 0}
+        const bom = {skus: [], price_per_month: 0}
         return bom
     }
 }