+++ okitweb/static/okit/views/designer/js/resources/autonomous_database.js
@@ -44,7 +44,10 @@ class AutonomousDatabaseView extends OkitDesignerArtefactView {
     /*
     ** Property Sheet Load function
      */
-    loadProperties() {
+    newPropertiesSheet() {
+        this.properties_sheet = new AutonomousDatabaseProperties(this.artefact)
+    }
+    loadProperties1() {
         let okitJson = this.getOkitJson();
         let me = this;
         $(jqId(PROPERTIES_PANEL)).load("propertysheets/autonomous_database.html", () => {